fix syscall not available on darwin
