+++ runtime/utils.go
@@ -4,7 +4,6 @@ import (
 	"os"
 	"strconv"
 	"syscall"
-	"unsafe"
 )
 
 func NumberCompare(x, op, y string) bool {
@@ -177,24 +176,4 @@ func FileHasAPositiveSize(file string) bool {
 	}
 
 	return stat.Size > 0
-}
-
-func FileDescriptorIsTerminal(sm *StreamManager, fd string) bool {
-	original, err := sm.Get(fd)
-	if err != nil {
-		return false
-	}
-
-	file, ok := original.(interface{ Fd() uintptr })
-	if !ok {
-		return false
-	}
-
-	_, _, errno := syscall.Syscall(
-		syscall.SYS_IOCTL,
-		file.Fd(),
-		syscall.TCGETS,
-		uintptr(unsafe.Pointer(&syscall.Termios{})),
-	)
-	return errno == 0
 }
